requirements.txt
streamlit
openai
python-dotenv
pyngrok
